requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pyyaml>=5.4.1
